Refactored AnnotationBasedModelFactory
Changed class scanning to be narrowed mainly by a list of packages to
scan. This seems to about ~20% slower than excluding jar files, but with
the benefit of having no further maintainance.
Instead of detecting M class by name we do it by fetching X class direct
subclasses. This is about ~15% slower but has the benefit of not having
to annotate M classes as well.